Add a maroto script for setting up the environment first time.
But wait! It can be run at any time in the project. This is great for test new additions to it, or even to update the whole dev environment if you need it.

You can use the flag -s when running it to skip the database creation and migration. Just because this can take time. If you know that you databases are up to date, feel free to use this flag.