Refactor: `like!` :point_right: `like`.
There wasn’t an ‘undestructive’ version of `like` before. So, no necessity of the bang pair.

And even using the Rail’s semantics, in this case, the exception raised by the `save!` method wasn’t being used anywhere.

And for finals: the vote returned by the method was just being used inside a spec. So, let’s fallback to the good’ol `save` and verifying for its truthiness inside the controllers and such.